pyproject.toml
[tool.poetry]
name = "djangoproject2"
version = "0.1.0"
description = ""
authors = ["islom <eshonqulovislom1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.6"
django-filter = "^23.2"
coreapi = "^2.3.3"
pillow = "^10.0.0"
pdf2image = "^1.16.3"
pypdf4 = "^1.27.0"
fitz = "^0.0.1.dev2"
frontend = "^0.0.3"
tools = "^0.1.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
